moved db_query to a seperate utils package
